Learning

+++ StudentManagementSystem/src/main/java/com/example/StudentManagementSystem/Model/Student.java
@@ -9,22 +9,22 @@ import jakarta.persistence.Id;
 public class Student {
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private int id;
+	private Long id;
 	private String name;
 	private int age;
 	
-	public Student(int id, String name, int age) {
+	public Student(Long id, String name, int age) {
 		this.id = id;
 		this.name = name;
 		this.age = age;
 		
 	}
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
